requirements.txt
requests>=2.32.3
selenium>=4.4.0
PyYAML>=6.0.1
